Extension exercise #1 done
Because we are already using the exit function, we can simply change
from 5 iterations to an infinite loop. Now the only way out of that loop
is to get it right... We probably need another way for the user to get
out in case the game is too hard for them - stay tuned!